requirements.txt
streamlit
pandas
requests
streamlit-oauth
PyJWT